// frontend/src/app/model/Acheteur.ts
export class Acheteur {
	achet_id: Number;
	achet_numero: Number;
	achet_numeroclient: Number;
	achet_numeroaffaire: Number;
	achet_type: String;
	achet_chiffreaffaire: Number;
	achet_incidentpaiement: Number;
	achet_montantincidentpaiement: Number;
	achet_montantpaiementrecup: Number;
	achet_dispersion: Number;
	achet_qualite: Number;
	achet_typologie: String;
	achet_creditencours: String;
	achet_montantcredit: Number;
	achet_montantrembours: Number;
	achet_montantecheance: Number;
	achet_montantecheancecredit: Number;
	achet_montantecheanceimpaye: Number;
	achet_montantimpaye: Number;
	achet_montantrecouvre: Number;
	achet_codeutilisateur: String;
	achet_datemodification: Date;

	achet_nom: String;
	achet_prenom: String;
	achet_comptebancaire: Number;
	achet_numclien_institu: String;
	achet_duree: Number;
	achet_avis: String;
	achet_date_avis: Date;
	achet_bon_commande: String;
	achet_date_facture: Date;
	achet_numero_facture: Number;
	achet_chiffreaffaire_confie: Number;
	achet_typecouverture: String;
	achet_bail: String;
	achet_duree_bail: Number;
	achet_montant_loyer: Number;
	achet_date_debut_credit: Date;

}